pyproject.toml
[tool.poetry]
name = "meet-tracker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["uvicorn"], version = "^0.115.0"}
SQLAlchemy = "^2.0.35"
asyncpg = "^0.29.0"
pydantic = {extras = ["email"], version = "^2.9.2"}
alembic = "^1.13.3"
psycopg2 = "^2.9.9"
passlib = "^1.7.4"
bcrypt = "4.0.1"
pyjwt = "^2.9.0"
python-multipart = "^0.0.12"
fastapi-cors = "^0.0.6"
pydub = "^0.25.1"
orjson = "^3.10.11"
faster-whisper = "^1.0.3"
numpy = "2.0.0"
torch = "^2.5.1"
aiobotocore = "^2.15.2"
aiofiles = "^24.1.0"
langchain-ollama = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
